ran prettier plugin through contracts

+++ contracts/Aqueduct.sol
@@ -4,33 +4,36 @@ import "./RBCN.sol";
 
 // This contract aims to serve as a holder of the community's RBCN
 contract Aqueduct {
-
     address public RBCNAddress;
 
     address public RubiconMarketAddress;
 
-    uint public distributionDuration;
+    uint256 public distributionDuration;
 
     address public owner;
-    
+
     /// @notice The address of the Rubicon governance token
     RBCNInterface public RBCN;
 
-    uint public timeOfLastRBCNDist;
+    uint256 public timeOfLastRBCNDist;
 
-    uint public propToMakers;     // the number out of 100 that represents proportion of an RBCN trade distribution to go to Maker vs. Taker
+    uint256 public propToMakers; // the number out of 100 that represents proportion of an RBCN trade distribution to go to Maker vs. Taker
 
-    event Distribution( address recipient, uint amount, uint timestamp);
+    event Distribution(address recipient, uint256 amount, uint256 timestamp);
 
-    constructor(uint communityDuration, address _owner) public {
-        require(communityDuration !=  0);
+    constructor(uint256 communityDuration, address _owner) public {
+        require(communityDuration != 0);
         owner = _owner;
         distributionDuration = communityDuration;
         // Initial 60/40 distribution of RBCN between makers and takers
         propToMakers = 60;
     }
-    
-    function setDistributionParams(address _RBCNAddress, address RubiconMarket) public onlyOwner returns (bool) {
+
+    function setDistributionParams(address _RBCNAddress, address RubiconMarket)
+        public
+        onlyOwner
+        returns (bool)
+    {
         RBCNAddress = _RBCNAddress;
         RubiconMarketAddress = RubiconMarket;
         RBCN = RBCNInterface(RBCNAddress);
@@ -63,7 +66,7 @@ contract Aqueduct {
     /**
      * @dev Safe subtraction function
      */
-    function sub(uint x, uint y) internal pure returns (uint z) {
+    function sub(uint256 x, uint256 y) internal pure returns (uint256 z) {
         require((z = x - y) <= x, "ds-math-sub-underflow");
     }
 
@@ -77,72 +80,111 @@ contract Aqueduct {
     /**
      * @dev Admin has the ability to choose a new exchange.
      */
-    function setNewExchange(address newImplementation) public onlyOwner returns (bool) {
+    function setNewExchange(address newImplementation)
+        public
+        onlyOwner
+        returns (bool)
+    {
         RubiconMarketAddress = newImplementation;
-    } 
+    }
 
     /**
      * @dev Only the owner can set a new owner. Timelock can be set as owner
      */
     function setOwner(address newOwner) public onlyOwner returns (bool) {
         owner = newOwner;
-    } 
+    }
 
     function getRBCNAddress() public view returns (address) {
-      return RBCNAddress;
+        return RBCNAddress;
     }
 
     // Should return the proportion of RBCN distribution per block that
     // is allocated to a Maker of a trade
     // The allocation that goes to the Taker is 1 - % to Maker
-    function getPropToMakers() public view returns(uint) {
-      return propToMakers;
+    function getPropToMakers() public view returns (uint256) {
+        return propToMakers;
     }
 
-    function setPropToMakers(uint newProp) external onlyOwner returns (bool){
-      propToMakers = newProp;
-      return true;
+    function setPropToMakers(uint256 newProp)
+        external
+        onlyOwner
+        returns (bool)
+    {
+        propToMakers = newProp;
+        return true;
     }
 
-    function distributeGovernanceToken(address recipient, uint amount) internal returns (bool) {
-        require(msg.sender == RubiconMarketAddress, "caller is not Rubicon Market");
+    function distributeGovernanceToken(address recipient, uint256 amount)
+        internal
+        returns (bool)
+    {
+        require(
+            msg.sender == RubiconMarketAddress,
+            "caller is not Rubicon Market"
+        );
         require(RBCN.transfer(recipient, amount), "transfer of RBCN failed");
         emit Distribution(recipient, amount, block.timestamp);
         return true;
     }
-    
+
     //This function should distribute a time-weighted RBCN allocation
-    function distributeToMakerAndTaker(address maker, address taker) public onlyExchange returns (bool) {
-      require(timeOfLastRBCNDist < block.timestamp, "timeOfLastRBCNDist < block.timestamp");
-      require(taker != address(0), "taker is zero address");
-      require(maker != address(0), "maker is zero address");
-      require(block.timestamp <= RBCN.getDistEndTime(), "RBCN Distribution is over");
+    function distributeToMakerAndTaker(address maker, address taker)
+        public
+        onlyExchange
+        returns (bool)
+    {
+        require(
+            timeOfLastRBCNDist < block.timestamp,
+            "timeOfLastRBCNDist < block.timestamp"
+        );
+        require(taker != address(0), "taker is zero address");
+        require(maker != address(0), "maker is zero address");
+        require(
+            block.timestamp <= RBCN.getDistEndTime(),
+            "RBCN Distribution is over"
+        );
 
-      //calculate change in time from last distribution to now
-      uint delta = sub(block.timestamp, timeOfLastRBCNDist);
+        //calculate change in time from last distribution to now
+        uint256 delta = sub(block.timestamp, timeOfLastRBCNDist);
 
-      //calculate quantity to send maker and taker
-      uint distQuanityMaker = (getPropToMakers() * (delta) * RBCN.getDistRate()) / 100;
-      uint distQuanityTaker = ((100 - getPropToMakers())) * (delta) * RBCN.getDistRate() / 100;
+        //calculate quantity to send maker and taker
+        uint256 distQuanityMaker =
+            (getPropToMakers() * (delta) * RBCN.getDistRate()) / 100;
+        uint256 distQuanityTaker =
+            (((100 - getPropToMakers())) * (delta) * RBCN.getDistRate()) / 100;
 
-      // TO DO: Extrapolate everything to Aqueduct?
-      //send to Maker distQuanityMaker
-      require(distributeGovernanceToken(maker, distQuanityMaker), "distribution to maker failed");
+        // TO DO: Extrapolate everything to Aqueduct?
+        //send to Maker distQuanityMaker
+        require(
+            distributeGovernanceToken(maker, distQuanityMaker),
+            "distribution to maker failed"
+        );
 
-      //send to Taker distQuanityTaker
-      require(distributeGovernanceToken(taker, distQuanityTaker), "distribution to taker failed");
+        //send to Taker distQuanityTaker
+        require(
+            distributeGovernanceToken(taker, distQuanityTaker),
+            "distribution to taker failed"
+        );
 
-      //time of timeOfLastRBCNDist is set immediately after distribution is made
-      timeOfLastRBCNDist = block.timestamp;
+        //time of timeOfLastRBCNDist is set immediately after distribution is made
+        timeOfLastRBCNDist = block.timestamp;
 
-      return true;
+        return true;
     }
 }
 
 interface RBCNInterface {
-    function getPriorVotes(address account, uint blockNumber) external view returns (uint96);
-    function getDistRate() external pure returns (uint);
-    function getDistStartTime() external view returns (uint);
-    function getDistEndTime() external view returns (uint);
-    function transfer(address, uint) external returns (bool);
+    function getPriorVotes(address account, uint256 blockNumber)
+        external
+        view
+        returns (uint96);
+
+    function getDistRate() external pure returns (uint256);
+
+    function getDistStartTime() external view returns (uint256);
+
+    function getDistEndTime() external view returns (uint256);
+
+    function transfer(address, uint256) external returns (bool);
 }
