+++ contracts/RBCN.sol
@@ -1,8 +1,9 @@
 /**
  *Submitted for verification at Etherscan.io on 2020-03-04
-*/
+ */
 
 pragma solidity ^0.5.16;
+
 // pragma experimental ABIEncoderV2;
 
 contract RBCN {
@@ -16,33 +17,33 @@ contract RBCN {
     uint8 public constant decimals = 18;
 
     /// @notice Total number of tokens in circulation
-    uint public constant totalSupply = 1000000000e18; // 1 billion RBCN
+    uint256 public constant totalSupply = 1000000000e18; // 1 billion RBCN
 
     /// @notice The Unix Timestamp to begin RBCN public distribution
-    uint public distStartTime;
+    uint256 public distStartTime;
 
     /// @notice The Unix Timestamp to begin RBCN public distribution
-    uint public distEndTime;
-    
+    uint256 public distEndTime;
+
     /// @notice The contract that holds the RBCN public distribution
     address public aqueduct;
-    
+
     /// @ notice The rate of RBCN per unit of Unix time (millisecond) distributed
     ///          to the community
-    /// Selected Number: https://www.wolframalpha.com/input/?i=%281%2C000%2C000%2C000*%28.51%29%29%2F%28365*x*24*60*60%29+%3D+%284044409199048374306+%2F+1e18%29 
-    /// Rate calculation: https://www.wolframalpha.com/input/?i=%281%2C000%2C000%2C000*.51%29%2F1.261e%2B8 
+    /// Selected Number: https://www.wolframalpha.com/input/?i=%281%2C000%2C000%2C000*%28.51%29%29%2F%28365*x*24*60*60%29+%3D+%284044409199048374306+%2F+1e18%29
+    /// Rate calculation: https://www.wolframalpha.com/input/?i=%281%2C000%2C000%2C000*.51%29%2F1.261e%2B8
     /// Rate = 4044409199048374 / 1e18 = 4.044409199...
-    uint public constant distRate = 4044409199048374306; //Distribution rate Wad in SECONDS
+    uint256 public constant distRate = 4044409199048374306; //Distribution rate Wad in SECONDS
     /// True rate of RBCN per second is distRate / 1e18
 
     /// @notice Allowance amounts on behalf of others
-    mapping (address => mapping (address => uint96)) internal allowances;
+    mapping(address => mapping(address => uint96)) internal allowances;
 
     /// @notice Official record of token balances for each account
-    mapping (address => uint96) internal balances;
+    mapping(address => uint96) internal balances;
 
     /// @notice A record of each accounts delegate
-    mapping (address => address) public delegates;
+    mapping(address => address) public delegates;
 
     /// @notice A checkpoint for marking number of votes from a given block
     struct Checkpoint {
@@ -51,31 +52,47 @@ contract RBCN {
     }
 
     /// @notice A record of votes checkpoints for each account, by index
-    mapping (address => mapping (uint32 => Checkpoint)) public checkpoints;
+    mapping(address => mapping(uint32 => Checkpoint)) public checkpoints;
 
     /// @notice The number of checkpoints for each account
-    mapping (address => uint32) public numCheckpoints;
+    mapping(address => uint32) public numCheckpoints;
 
     /// @notice The EIP-712 typehash for the contract's domain
-    bytes32 public constant DOMAIN_TYPEHASH = keccak256("EIP712Domain(string name,uint256 chainId,address verifyingContract)");
+    bytes32 public constant DOMAIN_TYPEHASH =
+        keccak256(
+            "EIP712Domain(string name,uint256 chainId,address verifyingContract)"
+        );
 
     /// @notice The EIP-712 typehash for the delegation struct used by the contract
-    bytes32 public constant DELEGATION_TYPEHASH = keccak256("Delegation(address delegatee,uint256 nonce,uint256 expiry)");
+    bytes32 public constant DELEGATION_TYPEHASH =
+        keccak256("Delegation(address delegatee,uint256 nonce,uint256 expiry)");
 
     /// @notice A record of states for signing / validating signatures
-    mapping (address => uint) public nonces;
+    mapping(address => uint256) public nonces;
 
     /// @notice An event thats emitted when an account changes its delegate
-    event DelegateChanged(address indexed delegator, address indexed fromDelegate, address indexed toDelegate);
+    event DelegateChanged(
+        address indexed delegator,
+        address indexed fromDelegate,
+        address indexed toDelegate
+    );
 
     /// @notice An event thats emitted when a delegate account's vote balance changes
-    event DelegateVotesChanged(address indexed delegate, uint previousBalance, uint newBalance);
+    event DelegateVotesChanged(
+        address indexed delegate,
+        uint256 previousBalance,
+        uint256 newBalance
+    );
 
     /// @notice The standard EIP-20 transfer event
     event Transfer(address indexed from, address indexed to, uint256 amount);
 
     /// @notice The standard EIP-20 approval event
-    event Approval(address indexed owner, address indexed spender, uint256 amount);
+    event Approval(
+        address indexed owner,
+        address indexed spender,
+        uint256 amount
+    );
 
     /**
      * @notice Construct a new RBCN token
@@ -85,8 +102,8 @@ contract RBCN {
     constructor(address aqueduct, address admin) public {
         // 51% to community
         balances[aqueduct] = uint96(510000000e18);
-        emit Transfer(address(0), aqueduct, uint(1000000000e18));
-        
+        emit Transfer(address(0), aqueduct, uint256(1000000000e18));
+
         // 49% to admin. See distribution here:
         balances[admin] = uint96(490000000e18);
         emit Transfer(address(0), admin, uint96(490000000e18));
@@ -102,7 +119,11 @@ contract RBCN {
      * @param spender The address of the account spending the funds
      * @return The number of tokens approved
      */
-    function allowance(address account, address spender) external view returns (uint) {
+    function allowance(address account, address spender)
+        external
+        view
+        returns (uint256)
+    {
         return allowances[account][spender];
     }
 
@@ -114,9 +135,12 @@ contract RBCN {
      * @param rawAmount The number of tokens that are approved (2^256-1 means infinite)
      * @return Whether or not the approval succeeded
      */
-    function approve(address spender, uint rawAmount) external returns (bool) {
+    function approve(address spender, uint256 rawAmount)
+        external
+        returns (bool)
+    {
         uint96 amount;
-        if (rawAmount == uint(-1)) {
+        if (rawAmount == uint256(-1)) {
             amount = uint96(-1);
         } else {
             amount = safe96(rawAmount, "RBCN::approve: amount exceeds 96 bits");
@@ -133,7 +157,7 @@ contract RBCN {
      * @param account The address of the account to get the balance of
      * @return The number of tokens held
      */
-    function balanceOf(address account) external view returns (uint) {
+    function balanceOf(address account) external view returns (uint256) {
         return balances[account];
     }
 
@@ -143,8 +167,9 @@ contract RBCN {
      * @param rawAmount The number of tokens to transfer
      * @return Whether or not the transfer succeeded
      */
-    function transfer(address dst, uint rawAmount) external returns (bool) {
-        uint96 amount = safe96(rawAmount, "RBCN::transfer: amount exceeds 96 bits");
+    function transfer(address dst, uint256 rawAmount) external returns (bool) {
+        uint96 amount =
+            safe96(rawAmount, "RBCN::transfer: amount exceeds 96 bits");
         _transferTokens(msg.sender, dst, amount);
         return true;
     }
@@ -156,13 +181,23 @@ contract RBCN {
      * @param rawAmount The number of tokens to transfer
      * @return Whether or not the transfer succeeded
      */
-    function transferFrom(address src, address dst, uint rawAmount) external returns (bool) {
+    function transferFrom(
+        address src,
+        address dst,
+        uint256 rawAmount
+    ) external returns (bool) {
         address spender = msg.sender;
         uint96 spenderAllowance = allowances[src][spender];
-        uint96 amount = safe96(rawAmount, "RBCN::approve: amount exceeds 96 bits");
+        uint96 amount =
+            safe96(rawAmount, "RBCN::approve: amount exceeds 96 bits");
 
         if (spender != src && spenderAllowance != uint96(-1)) {
-            uint96 newAllowance = sub96(spenderAllowance, amount, "RBCN::transferFrom: transfer amount exceeds spender allowance");
+            uint96 newAllowance =
+                sub96(
+                    spenderAllowance,
+                    amount,
+                    "RBCN::transferFrom: transfer amount exceeds spender allowance"
+                );
             allowances[src][spender] = newAllowance;
 
             emit Approval(src, spender, newAllowance);
@@ -189,13 +224,40 @@ contract RBCN {
      * @param r Half of the ECDSA signature pair
      * @param s Half of the ECDSA signature pair
      */
-    function delegateBySig(address delegatee, uint nonce, uint expiry, uint8 v, bytes32 r, bytes32 s) public {
-        bytes32 domainSeparator = keccak256(abi.encode(DOMAIN_TYPEHASH, keccak256(bytes(name)), getChainId(), address(this)));
-        bytes32 structHash = keccak256(abi.encode(DELEGATION_TYPEHASH, delegatee, nonce, expiry));
-        bytes32 digest = keccak256(abi.encodePacked("\x19\x01", domainSeparator, structHash));
+    function delegateBySig(
+        address delegatee,
+        uint256 nonce,
+        uint256 expiry,
+        uint8 v,
+        bytes32 r,
+        bytes32 s
+    ) public {
+        bytes32 domainSeparator =
+            keccak256(
+                abi.encode(
+                    DOMAIN_TYPEHASH,
+                    keccak256(bytes(name)),
+                    getChainId(),
+                    address(this)
+                )
+            );
+        bytes32 structHash =
+            keccak256(
+                abi.encode(DELEGATION_TYPEHASH, delegatee, nonce, expiry)
+            );
+        bytes32 digest =
+            keccak256(
+                abi.encodePacked("\x19\x01", domainSeparator, structHash)
+            );
         address signatory = ecrecover(digest, v, r, s);
-        require(signatory != address(0), "RBCN::delegateBySig: invalid signature");
-        require(nonce == nonces[signatory]++, "RBCN::delegateBySig: invalid nonce");
+        require(
+            signatory != address(0),
+            "RBCN::delegateBySig: invalid signature"
+        );
+        require(
+            nonce == nonces[signatory]++,
+            "RBCN::delegateBySig: invalid nonce"
+        );
         require(now <= expiry, "RBCN::delegateBySig: signature expired");
         return _delegate(signatory, delegatee);
     }
@@ -207,7 +269,8 @@ contract RBCN {
      */
     function getCurrentVotes(address account) external view returns (uint96) {
         uint32 nCheckpoints = numCheckpoints[account];
-        return nCheckpoints > 0 ? checkpoints[account][nCheckpoints - 1].votes : 0;
+        return
+            nCheckpoints > 0 ? checkpoints[account][nCheckpoints - 1].votes : 0;
     }
 
     /**
@@ -217,8 +280,15 @@ contract RBCN {
      * @param blockNumber The block number to get the vote balance at
      * @return The number of votes the account had as of the given block
      */
-    function getPriorVotes(address account, uint blockNumber) public view returns (uint96) {
-        require(blockNumber < block.number, "RBCN::getPriorVotes: not yet determined");
+    function getPriorVotes(address account, uint256 blockNumber)
+        public
+        view
+        returns (uint96)
+    {
+        require(
+            blockNumber < block.number,
+            "RBCN::getPriorVotes: not yet determined"
+        );
 
         uint32 nCheckpoints = numCheckpoints[account];
         if (nCheckpoints == 0) {
@@ -251,16 +321,16 @@ contract RBCN {
         return checkpoints[account][lower].votes;
     }
 
-    function getDistStartTime() external view returns (uint) {
-      return distStartTime;
+    function getDistStartTime() external view returns (uint256) {
+        return distStartTime;
     }
 
-    function getDistEndTime() external view returns (uint) {
-      return distEndTime;
+    function getDistEndTime() external view returns (uint256) {
+        return distEndTime;
     }
 
-    function getDistRate() external pure returns (uint) {
-      return distRate;
+    function getDistRate() external pure returns (uint256) {
+        return distRate;
     }
 
     function _delegate(address delegator, address delegatee) internal {
@@ -273,72 +343,143 @@ contract RBCN {
         _moveDelegates(currentDelegate, delegatee, delegatorBalance);
     }
 
-    function _transferTokens(address src, address dst, uint96 amount) internal {
-        require(src != address(0), "RBCN::_transferTokens: cannot transfer from the zero address");
-        require(dst != address(0), "RBCN::_transferTokens: cannot transfer to the zero address");
+    function _transferTokens(
+        address src,
+        address dst,
+        uint96 amount
+    ) internal {
+        require(
+            src != address(0),
+            "RBCN::_transferTokens: cannot transfer from the zero address"
+        );
+        require(
+            dst != address(0),
+            "RBCN::_transferTokens: cannot transfer to the zero address"
+        );
 
-        balances[src] = sub96(balances[src], amount, "RBCN::_transferTokens: transfer amount exceeds balance");
-        balances[dst] = add96(balances[dst], amount, "RBCN::_transferTokens: transfer amount overflows");
+        balances[src] = sub96(
+            balances[src],
+            amount,
+            "RBCN::_transferTokens: transfer amount exceeds balance"
+        );
+        balances[dst] = add96(
+            balances[dst],
+            amount,
+            "RBCN::_transferTokens: transfer amount overflows"
+        );
         emit Transfer(src, dst, amount);
 
         _moveDelegates(delegates[src], delegates[dst], amount);
     }
 
-    function _moveDelegates(address srcRep, address dstRep, uint96 amount) internal {
+    function _moveDelegates(
+        address srcRep,
+        address dstRep,
+        uint96 amount
+    ) internal {
         if (srcRep != dstRep && amount > 0) {
             if (srcRep != address(0)) {
                 uint32 srcRepNum = numCheckpoints[srcRep];
-                uint96 srcRepOld = srcRepNum > 0 ? checkpoints[srcRep][srcRepNum - 1].votes : 0;
-                uint96 srcRepNew = sub96(srcRepOld, amount, "RBCN::_moveVotes: vote amount underflows");
+                uint96 srcRepOld =
+                    srcRepNum > 0
+                        ? checkpoints[srcRep][srcRepNum - 1].votes
+                        : 0;
+                uint96 srcRepNew =
+                    sub96(
+                        srcRepOld,
+                        amount,
+                        "RBCN::_moveVotes: vote amount underflows"
+                    );
                 _writeCheckpoint(srcRep, srcRepNum, srcRepOld, srcRepNew);
             }
 
             if (dstRep != address(0)) {
                 uint32 dstRepNum = numCheckpoints[dstRep];
-                uint96 dstRepOld = dstRepNum > 0 ? checkpoints[dstRep][dstRepNum - 1].votes : 0;
-                uint96 dstRepNew = add96(dstRepOld, amount, "RBCN::_moveVotes: vote amount overflows");
+                uint96 dstRepOld =
+                    dstRepNum > 0
+                        ? checkpoints[dstRep][dstRepNum - 1].votes
+                        : 0;
+                uint96 dstRepNew =
+                    add96(
+                        dstRepOld,
+                        amount,
+                        "RBCN::_moveVotes: vote amount overflows"
+                    );
                 _writeCheckpoint(dstRep, dstRepNum, dstRepOld, dstRepNew);
             }
         }
     }
 
-    function _writeCheckpoint(address delegatee, uint32 nCheckpoints, uint96 oldVotes, uint96 newVotes) internal {
-      uint32 blockNumber = safe32(block.number, "RBCN::_writeCheckpoint: block number exceeds 32 bits");
+    function _writeCheckpoint(
+        address delegatee,
+        uint32 nCheckpoints,
+        uint96 oldVotes,
+        uint96 newVotes
+    ) internal {
+        uint32 blockNumber =
+            safe32(
+                block.number,
+                "RBCN::_writeCheckpoint: block number exceeds 32 bits"
+            );
 
-      if (nCheckpoints > 0 && checkpoints[delegatee][nCheckpoints - 1].fromBlock == blockNumber) {
-          checkpoints[delegatee][nCheckpoints - 1].votes = newVotes;
-      } else {
-          checkpoints[delegatee][nCheckpoints] = Checkpoint(blockNumber, newVotes);
-          numCheckpoints[delegatee] = nCheckpoints + 1;
-      }
+        if (
+            nCheckpoints > 0 &&
+            checkpoints[delegatee][nCheckpoints - 1].fromBlock == blockNumber
+        ) {
+            checkpoints[delegatee][nCheckpoints - 1].votes = newVotes;
+        } else {
+            checkpoints[delegatee][nCheckpoints] = Checkpoint(
+                blockNumber,
+                newVotes
+            );
+            numCheckpoints[delegatee] = nCheckpoints + 1;
+        }
 
-      emit DelegateVotesChanged(delegatee, oldVotes, newVotes);
+        emit DelegateVotesChanged(delegatee, oldVotes, newVotes);
     }
 
-    function safe32(uint n, string memory errorMessage) internal pure returns (uint32) {
+    function safe32(uint256 n, string memory errorMessage)
+        internal
+        pure
+        returns (uint32)
+    {
         require(n < 2**32, errorMessage);
         return uint32(n);
     }
 
-    function safe96(uint n, string memory errorMessage) internal pure returns (uint96) {
+    function safe96(uint256 n, string memory errorMessage)
+        internal
+        pure
+        returns (uint96)
+    {
         require(n < 2**96, errorMessage);
         return uint96(n);
     }
 
-    function add96(uint96 a, uint96 b, string memory errorMessage) internal pure returns (uint96) {
+    function add96(
+        uint96 a,
+        uint96 b,
+        string memory errorMessage
+    ) internal pure returns (uint96) {
         uint96 c = a + b;
         require(c >= a, errorMessage);
         return c;
     }
 
-    function sub96(uint96 a, uint96 b, string memory errorMessage) internal pure returns (uint96) {
+    function sub96(
+        uint96 a,
+        uint96 b,
+        string memory errorMessage
+    ) internal pure returns (uint96) {
         require(b <= a, errorMessage);
         return a - b;
     }
 
-    function getChainId() internal pure returns (uint) {
+    function getChainId() internal pure returns (uint256) {
         uint256 chainId;
-        assembly { chainId := chainid() }
+        assembly {
+            chainId := chainid()
+        }
         return chainId;
     }
 }
